WIP show image hotspot removal details
