+++ metaspace/webapp/package.json
@@ -19,6 +19,7 @@
     "@types/js-cookie": "^2.1.0",
     "@types/marked": "^0.6.2",
     "@types/sanitize-html": "^1.18.2",
+    "@types/upng-js": "^2.1.1",
     "@types/ws": "^5.1.2",
     "apollo-client": "^2.3.7",
     "apollo-client-preset": "^1.0.8",
@@ -61,9 +62,10 @@
     "sanitize-html": "^1.20.0",
     "sass-loader": "7.0.1",
     "serve-favicon": "^2.3.2",
-    "simple-statistics": "^3.0.0",
+    "simple-statistics": "^7.0.2",
     "subscriptions-transport-ws": "^0.9.14",
     "ts-loader": "^4.3.0",
+    "upng-js": "^2.1.0",
     "uuid": "^3.0.1",
     "vue": "~2.5.16",
     "vue-analytics": "^5.9.1",
