+++ metaspace/webapp/src/config.ts
@@ -33,6 +33,7 @@ interface Features {
 interface ClientConfig {
   graphqlUrl: string | null;
   wsGraphqlUrl: string | null;
+  imageStorage?: string | null;
 
   google_client_id: string;
